update(10)

+++ database/migrations/2021_08_05_075149_create_product_ratings_table.php
@@ -15,7 +15,7 @@ class CreateProductRatingsTable extends Migration
     {
         Schema::create('product_ratings', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('rating');
+            $table->float('rating');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
             $table->timestamps();
